[METRIC] Move serializer out from producer package (#1647)
