+++ common/src/main/java/org/astraea/common/backup/RecordReaderBuilder.java
@@ -25,6 +25,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.function.Function;
 import java.util.zip.GZIPInputStream;
+import org.astraea.common.ByteUtils;
 import org.astraea.common.Header;
 import org.astraea.common.Utils;
 import org.astraea.common.consumer.Record;
