+++ common/src/main/java/org/astraea/common/producer/Serializer.java
@@ -28,14 +28,13 @@ import org.apache.kafka.common.serialization.FloatSerializer;
 import org.apache.kafka.common.serialization.IntegerSerializer;
 import org.apache.kafka.common.serialization.LongSerializer;
 import org.apache.kafka.common.serialization.StringSerializer;
+import org.astraea.common.ByteUtils;
 import org.astraea.common.Header;
 import org.astraea.common.Utils;
 import org.astraea.common.admin.ClusterInfo;
 import org.astraea.common.admin.NodeInfo;
 import org.astraea.common.admin.Replica;
 import org.astraea.common.admin.Topic;
-import org.astraea.common.backup.ByteUtils;
-import org.astraea.common.generated.BeanObjectOuterClass;
 import org.astraea.common.json.JsonConverter;
 import org.astraea.common.json.TypeRef;
 import org.astraea.common.metrics.BeanObject;
@@ -84,6 +83,7 @@ public interface Serializer<T> {
   Serializer<Topic> TOPIC = new TopicSerializer();
   Serializer<Replica> REPLICA = new ReplicaSerializer();
   Serializer<ClusterInfo> CLUSTER_INFO = new ClusterInfoSerializer();
+  Serializer<BeanObject> BEAN_OBJECT = (topic, headers, data) -> ByteUtils.toBytes(data);
 
   /**
    * create Custom JsonSerializer
@@ -249,40 +249,6 @@ public interface Serializer<T> {
             buffer.put(bytes);
           });
       return buffer.array();
-    }
-  }
-
-  class BeanObjectSerializer implements Serializer<BeanObject> {
-    @Override
-    public byte[] serialize(String topic, Collection<Header> headers, BeanObject data) {
-      var beanBuilder = BeanObjectOuterClass.BeanObject.newBuilder();
-      beanBuilder.setDomain(data.domainName());
-      beanBuilder.putAllProperties(data.properties());
-      data.attributes().forEach((key, val) -> beanBuilder.putAttributes(key, primitive(val)));
-
-      return beanBuilder.build().toByteArray();
-    }
-
-    private static BeanObjectOuterClass.BeanObject.Primitive primitive(Object v) {
-      if (v instanceof Integer)
-        return BeanObjectOuterClass.BeanObject.Primitive.newBuilder().setInt((int) v).build();
-      else if (v instanceof Long)
-        return BeanObjectOuterClass.BeanObject.Primitive.newBuilder().setLong((long) v).build();
-      else if (v instanceof Float)
-        return BeanObjectOuterClass.BeanObject.Primitive.newBuilder().setFloat((float) v).build();
-      else if (v instanceof Double)
-        return BeanObjectOuterClass.BeanObject.Primitive.newBuilder().setDouble((double) v).build();
-      else if (v instanceof Boolean)
-        return BeanObjectOuterClass.BeanObject.Primitive.newBuilder()
-            .setBoolean((boolean) v)
-            .build();
-      else if (v instanceof String)
-        return BeanObjectOuterClass.BeanObject.Primitive.newBuilder().setStr(v.toString()).build();
-      else
-        throw new IllegalArgumentException(
-            "Type "
-                + v.getClass()
-                + " is not supported. Please use Integer, Long, Float, Double, Boolean, String instead.");
     }
   }
 }
