+++ common/src/test/java/org/astraea/common/metrics/collector/MetricFetcherTest.java
@@ -157,13 +157,16 @@ public class MetricFetcherTest {
       try (var consumer =
           Consumer.forTopics(Set.of("__metrics"))
               .bootstrapServers(SERVICE.bootstrapServers())
-              .valueDeserializer(Deserializer.STRING)
+              .valueDeserializer(Deserializer.BEAN_OBJECT)
               .seek(SeekStrategy.DISTANCE_FROM_BEGINNING, 0)
               .build()) {
         var records =
             consumer.poll(Duration.ofSeconds(5)).stream().collect(Collectors.toUnmodifiableList());
         Assertions.assertEquals(1, records.size());
-        Assertions.assertEquals(testBean.toString(), records.get(0).value());
+        var getBean = records.get(0).value();
+        Assertions.assertEquals(testBean.domainName(), getBean.domainName());
+        Assertions.assertEquals(testBean.properties(), getBean.properties());
+        Assertions.assertEquals(testBean.attributes(), getBean.attributes());
       }
     }
   }
