+++ common/src/test/java/org/astraea/common/serializer/ClusterInfoSerializerTest.java
@@ -23,12 +23,12 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import org.apache.kafka.common.Node;
+import org.astraea.common.ByteUtils;
 import org.astraea.common.Utils;
 import org.astraea.common.admin.Admin;
 import org.astraea.common.admin.ClusterInfo;
 import org.astraea.common.admin.NodeInfo;
 import org.astraea.common.admin.TopicPartition;
-import org.astraea.common.backup.ByteUtils;
 import org.astraea.common.consumer.Deserializer;
 import org.astraea.common.producer.Serializer;
 import org.astraea.it.Service;
